restructure variable resolver
